# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing drivers space",
        "title": "DriversSpace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Selmi hamza",
            "email": "hamza@craftfoundry.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api-company.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/company": {
            "get": {
                "description": "retrieve a company by ID in database with the ID in request parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver_companies"
                ],
                "summary": "retrieve a company by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter id:",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.DriverCompanie"
                        }
                    }
                }
            },
            "put": {
                "description": "update a company by id with the input payload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver_companies"
                ],
                "summary": "update a company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter id:",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyName:",
                        "name": "CompanyName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyDre:",
                        "name": "PcoNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyTva:",
                        "name": "VatNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanySiren:",
                        "name": "Registration",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "enter VatPercentage:",
                        "name": "VatRate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.DriverCompanie"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new company with the input payload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver_companies"
                ],
                "summary": "Create a new company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter driver id (tester avec l'id : 17357 ) : ",
                        "name": "DriverID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyName:",
                        "name": "CompanyName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyDre:",
                        "name": "PcoNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanyTva:",
                        "name": "VatNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enter CompanySiren:",
                        "name": "Registration",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "enter VatPercentage:",
                        "name": "VatRate",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "enter rib file:",
                        "name": "RIB",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "enter insee file:",
                        "name": "INSEE",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "enter dre file:",
                        "name": "DRE",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.DriverCompanie"
                        }
                    }
                }
            }
        },
...
        "/request/{DriverID}": {
            "get": {
                "description": "retrieve a request in database with the driver ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request"
                ],
                "summary": "retrieve a request by driver ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter driver id :11469:",
                        "name": "DriverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.Request"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Models.DriverCompanie": {
            "type": "object",
            "properties": {
                "companyDre": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "companySiren": {
                    "type": "string"
                },
                "companyTva": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Models.FileCompanie"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vatPercentage": {
                    "type": "number"
                }
            }
        },
        "Models.FileCompanie": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "AdministrativeFileType AdministrativeFileType `gorm:\"foreignkey:FileType;association_foreignkey:ID\"`",
                    "type": "string"
                },
                "driverCompanyID": {
                    "type": "integer"
                },
                "fileType": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ref": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Models.Request": {
            "type": "object",
            "properties": {
                "activeForDispatch": {
                    "type": "boolean"
                },
                "addressDropOff": {
                    "type": "string"
                },
                "addressDropOffComplement": {
                    "type": "string"
                },
                "addressDropOffLat": {
                    "type": "number"
                },
                "addressDropOffLong": {
                    "type": "number"
                },
                "addressDropOffPostalCode": {
                    "type": "string"
                },
                "addressPickUp": {
                    "type": "string"
                },
                "addressPickUpComplement": {
                    "type": "string"
                },
                "addressPickUpLat": {
                    "type": "number"
                },
                "addressPickUpLong": {
                    "type": "number"
                },
                "addressPickUpPostalCode": {
                    "type": "string"
                },
                "agendaEventID": {
                    "type": "integer"
                },
                "channelID": {
                    "type": "integer"
                },
                "closed": {
                    "type": "boolean"
                },
                "comment": {
                    "type": "string"
                },
                "commercialFees": {
                    "type": "integer"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyPreferencesID": {
                    "type": "integer"
                },
                "costCenter": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "deliveryRecipientID": {
                    "description": "CompanyReferences",
                    "type": "integer"
                },
                "deliverySenderID": {
                    "type": "integer"
                },
                "directRequest": {
                    "type": "boolean"
                },
                "discountCodeDiscount": {
                    "type": "integer"
                },
                "discountCodeID": {
                    "type": "integer"
                },
                "dispatchDate": {
                    "type": "string"
                },
                "driverCarID": {
                    "type": "integer"
                },
                "driverDiscountCodeDiscount": {
                    "type": "integer"
                },
                "driverID": {
                    "type": "integer"
                },
                "driverPaymentID": {
                    "type": "integer"
                },
                "driverPrice": {
                    "type": "integer"
                },
                "driverZoneID": {
                    "description": "PriceComponents          JSONB",
                    "type": "integer"
                },
                "dropOffTimeWindowID": {
                    "type": "integer"
                },
                "estimateDistanceM": {
                    "type": "integer"
                },
                "estimateDropOffDate": {
                    "type": "string"
                },
                "estimatePickUpDate": {
                    "type": "string"
                },
                "fireTime": {
                    "type": "string"
                },
                "flightID": {
                    "type": "integer"
                },
                "flightNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "initialVehicleTypeID": {
                    "type": "integer"
                },
                "invoiceID": {
                    "type": "integer"
                },
                "maximumDetour": {
                    "type": "integer"
                },
                "onlyDriverPriceCents": {
                    "type": "integer"
                },
                "optionIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "optionsPrice": {
                    "type": "integer"
                },
                "outsourcedDestinationRequestID": {
                    "type": "integer"
                },
                "outsourcedOriginRequestID": {
                    "type": "integer"
                },
                "partnerID": {
                    "type": "string"
                },
                "passengerID": {
                    "type": "integer"
                },
                "payinID": {
                    "type": "integer"
                },
                "paymentType": {
                    "type": "string"
                },
                "pickUpTimeWindowID": {
                    "type": "integer"
                },
                "previousRequestID": {
                    "type": "integer"
                },
                "priceChanged": {
                    "type": "boolean"
                },
                "priorityRequest": {
                    "type": "boolean"
                },
                "recurrentRequestID": {
                    "type": "integer"
                },
                "requestDiscountCode": {
                    "type": "string"
                },
                "reservationCode": {
                    "type": "string"
                },
                "rideDate": {
                    "type": "string"
                },
                "rideDiscountPercentage": {
                    "type": "number"
                },
                "rideDropOffDate": {
                    "type": "string"
                },
                "ridePrice": {
                    "type": "number"
                },
                "rideType": {
                    "type": "integer"
                },
                "ruleValueTimeWindowID": {
                    "type": "integer"
                },
                "saasCompanyID": {
                    "type": "integer"
                },
                "saasOfficeID": {
                    "type": "integer"
                },
                "sidekiqJid": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "theoreticalDuration": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicleTypeID": {
                    "type": "integer"
                },
                "zoneID": {
                    "type": "integer"
                }
            }
        },
...
    }
}